Update NewLineAtEndOfFileRule
Update `NewLineAtEndOfFileRule`:
- Rename`NewLineAtTheEndRule` to `NewLineAtEndOfFileRule` to make
clear the rule should only be affected by the end of file.
- Change rule ID to `213`.
- Update code to make in simpler to read.
